+++ src/Data/ILevel1Data.cs
@@ -26,11 +26,11 @@ namespace TradeSharp.Data
     //properties
     public DateTime DateTime { get; set; }
     public double Bid { get; set; }
-    public long BidSize { get; set; }
+    public double BidSize { get; set; }
     public double Ask { get; set; }
-    public long AskSize { get; set; }
+    public double AskSize { get; set; }
     public double Last { get; set; }
-    public long LastSize { get; set; }
+    public double LastSize { get; set; }
 
     //methods
